requirements.txt

mlflow==2.10.2
pymilvus==2.4.5
protobuf==3.20.1
numpy==1.26.4
pandas
scipy
pyarrow
langchain==0.2.11
langchain-community
langchain-groq
python-dotenv
pydantic==1.10.13
-e .